+++ Biblioteca_auxiliar.h
@@ -9,7 +9,7 @@ void limpa();
 int verificaSenha(int tam, Cliente *clientes, char *compara);
 int verificaCPF(int tam, Cliente *clientes, char *compara);
 void arruma_tipo(int tamanho, char *tipo, Cliente *clientes);
-double aplica_taxa(int indice,  Cliente *clientes, int valor);
+double aplica_taxa(int indice,  Cliente *clientes, double valor);
 int confere_tipo(int indice, Cliente *clientes);
 
 #endif
